requirements.txt
fastapi
uvicorn[standard]
confluent-kafka
psycopg2-binary
requests
python-dotenv
pydantic
pydantic-settings
kafka-python